requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
clearbit==0.1.7
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.2
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2-binary==2.7.7
pyhunter==0.3
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
